requirements.txt
Flask
python-vlc
requests
comtypes
langchain_core 
langchain_community 
langchain_groq 
SpeechRecognition 
pyttsx3 
pyaudio 
yt-dlp 
gunicorn
py-espeak-ng